pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vector-db"
version = "0.1.0"
description = "A Vector Database engine built from scratch"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Georgio Bou Sleiman", email = "bousleimengeorgio139@gmail.com"},
]
keywords = ["vector-database", "embeddings", "similarity-search", "ann", "product-quantization"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "numba>=0.58.0",
    "pandas>=2.0.0",

    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",

    "pyyaml>=6.0.0",
    "locust>=2.0.0",
]

[project.optional-dependencies]
cpu = [
    "sentence-transformers>=2.2.0",
    "transformers>=4.30.0",
    "tokenizers>=0.13.0",
]
gpu = [
    "sentence-transformers>=2.2.0",
    "transformers>=4.30.0",
    "tokenizers>=0.13.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ipython>=8.12.0",
    "jupyter>=1.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "vector-db[dev]",
    "faiss-cpu>=1.7.4",
    "annoy>=1.17.3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vector_db"]

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PD",  # pandas-vet
    "PL",  # Pylint
    "TRY", # tryceratops
    "RUF", # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "PLR0913", # too many arguments
    "PLR2004", # magic value used in comparison
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["vector_db"]

[tool.ruff.mccabe]
max-complexity = 10

[tool.mypy]
python_version = "3.9"
# Less strict settings - only catch serious issues that affect development
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
disallow_untyped_decorators = false
no_implicit_optional = false
warn_redundant_casts = false
warn_unused_ignores = false
warn_no_return = false
warn_unreachable = false
strict_equality = false
show_error_codes = true
ignore_errors = false

[[tool.mypy.overrides]]
module = [
    "numba.*",
    "hnswlib.*",
    "faiss.*",
    "annoy.*",
    "scipy.*",
    "sentence_transformers.*",
    "yaml",
    "yaml.*",
    "torch",
    "torch.*",
    "pydantic",
    "pydantic.*",
    "fastapi",
    "fastapi.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=vector_db",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["vector_db"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

